requirements.txt
customtkinter
yt-dlp
packaging
requests
pillow
selenium
webdriver-manager
Flask
pywebview
